Moves apm echo version 4 into specific top-level module.
