+++ module/apmechov4/doc.go
@@ -19,4 +19,4 @@
 
 // Package apmecho provides middleware for the Echo framework,
 // for tracing HTTP requests.
-package apmecho
+package apmechov4
