+++ module/apmechov4/example_test.go
@@ -17,13 +17,13 @@
 
 // +build go1.9
 
-package apmecho_test
+package apmechov4_test
 
 import (
 	"github.com/labstack/echo/v4"
 
 	"go.elastic.co/apm"
-	"go.elastic.co/apm/module/apmecho/v4"
+	apmecho "go.elastic.co/apm/module/apmechov4"
 )
 
 func ExampleMiddleware() {
